Tiny refactor to remove out some methods that aren't needed. Also removed the about link which wasn't needed.
