+++ BlazorQuiz/Shared/MainLayout.razor
@@ -7,7 +7,6 @@
 
     <main>
         <div class="top-row px-4">
-            <a href="https://docs.microsoft.com/aspnet/" target="_blank">About</a>
         </div>
 
         <article class="content px-4">
